# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "在线聊天系统",
        "title": "chat",
        "contact": {},
        "license": {
            "name": "lyc,why"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8084",
    "basePath": "/",
    "paths": {
...
        "/api/application/accept": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "被申请者同意好友申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "需要同意的申请",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AcceptApplication"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2010:账号不存在 3002:申请不存在 3004:重复操作申请",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/application/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "发起好友申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "发起好友申请",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateApplicationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2001:鉴权失败 5003:好友已经存在 5004:不能添加自己为好友",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "删除已经发送的好友申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "删除已经发起的申请",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.DeleteApplicationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2001:鉴权失败 ",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/application/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "账户查看和自身相关的好友申请(不论是申请者还是被申请者)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "Page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "PageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1003:系统错误 2007:身份不存在 2008:身份验证失败 2010:账号不存在",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ApplicationsList"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/application/refuse": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "被申请者拒绝好友申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-token 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "需要拒绝的申请",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.RefuseApplication"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2010:账号不存在 3002:申请不存在 3004:重复操作申请",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.State": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "失败时返回空"
                },
                "status_code": {
                    "description": "状态码，0-成功，其他值-失败",
                    "type": "integer"
                },
                "status_msg": {
                    "description": "返回状态描述",
                    "type": "string"
                }
            }
        },
...
        "reply.ApplicationsList": {
            "type": "object",
            "properties": {
                "applicationList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reply.GetApplication"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "reply.EasyAccount": {
            "type": "object",
            "properties": {
                "accountID": {
                    "type": "integer"
                },
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "reply.GetApplication": {
            "type": "object",
            "properties": {
                "applicant": {
                    "$ref": "#/definitions/reply.EasyAccount"
                },
                "applyMsg": {
                    "type": "string"
                },
                "receiver": {
                    "$ref": "#/definitions/reply.EasyAccount"
                },
                "refuseMsg": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "request.AcceptApplication": {
            "type": "object",
            "required": [
                "applicant_id"
            ],
            "properties": {
                "applicant_id": {
                    "type": "integer"
                }
            }
        },
...
        "request.CreateApplicationReq": {
            "type": "object",
            "required": [
                "accountID",
                "applicationMsg"
            ],
            "properties": {
                "accountID": {
                    "type": "integer"
                },
                "applicationMsg": {
                    "type": "string"
                }
            }
        },
        "request.DeleteApplicationReq": {
            "type": "object",
            "required": [
                "accountID"
            ],
            "properties": {
                "accountID": {
                    "type": "integer"
                }
            }
        },
...
        "request.RefuseApplication": {
            "type": "object",
            "required": [
                "applicant_id"
            ],
            "properties": {
                "applicant_id": {
                    "type": "integer"
                },
                "refuse_msg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}